+++ AGauge/AGauge.cs
@@ -1435,427 +1435,7 @@ namespace System.Windows.Forms
     #endregion
   }
 
-  #region[ Gauge Range ]
-  public class AGaugeRangeCollection : CollectionBase
-  {
-    private AGauge Owner;
-    public AGaugeRangeCollection(AGauge sender) { Owner = sender; }
-    private void NotifyChanging()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanging();
-      }
-    }
-
-    private void NotifyChanged()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanged();
-      }
-    }
-
-    public AGaugeRange this[int index] { get { return (AGaugeRange)List[index]; } }
-    public bool Contains(AGaugeRange itemType) { return List.Contains(itemType); }
-    public int Add(AGaugeRange itemType)
-    {
-      itemType.SetOwner(Owner);
-      if (string.IsNullOrEmpty(itemType.Name)) itemType.Name = GetUniqueName();
-      var ret = List.Add(itemType);
-      if (Owner != null) Owner.RepaintControl();
-      return ret;
-    }
-    public void Remove(AGaugeRange itemType)
-    {
-      List.Remove(itemType);
-      if (Owner != null) Owner.RepaintControl();
-    }
-    public void Insert(int index, AGaugeRange itemType)
-    {
-      itemType.SetOwner(Owner);
-      if (string.IsNullOrEmpty(itemType.Name)) itemType.Name = GetUniqueName();
-      List.Insert(index, itemType);
-      if (Owner != null) Owner.RepaintControl();
-    }
-    public int IndexOf(AGaugeRange itemType) { return List.IndexOf(itemType); }
-    public AGaugeRange FindByName(string name)
-    {
-      foreach (AGaugeRange ptrRange in List)
-      {
-        if (ptrRange.Name == name) return ptrRange;
-      }
-      return null;
-    }
-
-    protected override void OnInsert(int index, object value)
-    {
-      NotifyChanging();
-      if (string.IsNullOrEmpty(((AGaugeRange)value).Name)) ((AGaugeRange)value).Name = GetUniqueName();
-      base.OnInsert(index, value);
-      ((AGaugeRange)value).SetOwner(Owner);
-      NotifyChanged();
-    }
-    protected override void OnRemove(int index, object value)
-    {
-      NotifyChanging();
-      if (Owner != null) Owner.RepaintControl();
-      NotifyChanged();
-    }
-    protected override void OnClear()
-    {
-      NotifyChanging();
-      if (Owner != null) Owner.RepaintControl();
-      NotifyChanged();
-    }
-
-    private string GetUniqueName()
-    {
-      const string Prefix = "GaugeRange";
-      int index = 1;
-      bool valid;
-      while (this.Count != 0)
-      {
-        valid = true;
-        for (int x = 0; x < this.Count; x++)
-        {
-          if (this[x].Name == (Prefix + index.ToString()))
-          {
-            valid = false;
-            break;
-          }
-        }
-        if (valid) break;
-        index++;
-      };
-      return Prefix + index.ToString();
-    }
-  }
-  public class AGaugeRange
-  {
-    public AGaugeRange() { }
-
-    public AGaugeRange(Color color, Single startValue, Single endValue)
-    {
-      Color = color;
-      _StartValue = startValue;
-      _EndValue = endValue;
-    }
-
-    public AGaugeRange(Color color, Single startValue, Single endValue, Int32 innerRadius, Int32 outerRadius)
-    {
-      Color = color;
-      _StartValue = startValue;
-      _EndValue = endValue;
-      InnerRadius = innerRadius;
-      OuterRadius = outerRadius;
-    }
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Design"),
-    System.ComponentModel.DisplayName("(Name)"),
-    System.ComponentModel.Description("Instance Name.")]
-    public string Name { get { return _Name; } set { NotifyChanging(); _Name = value; NotifyChanged(); } }
-    private string _Name; 
-
-    [System.ComponentModel.Browsable(false)]
-    public Boolean InRange { get; set; }
-
-    private AGauge Owner;
-    [System.ComponentModel.Browsable(false)]
-    public void SetOwner(AGauge value) { Owner = value; }
-    private void NotifyOwner() { if (Owner != null) Owner.RepaintControl(); }
-    private void NotifyChanging()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanging();
-      }
-    }
-
-    private void NotifyChanged()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanged();
-      }
-    }
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The color of the range.")]
-    public Color Color { get { return _Color; } set { NotifyChanging(); _Color = value; NotifyOwner(); NotifyChanged(); } }
-    private Color _Color;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Limits"),
-    System.ComponentModel.Description("The start value of the range, must be less than RangeEndValue.")]
-    public Single StartValue
-    {
-      get { return _StartValue; }
-      set
-      {
-        NotifyChanging();
-        if (Owner != null)
-        {
-          if (value < Owner.MinValue) value = Owner.MinValue;
-          if (value > Owner.MaxValue) value = Owner.MaxValue;
-        }
-        _StartValue = value;
-        NotifyOwner();
-        NotifyChanged();
-      }
-    }
-    private Single _StartValue;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Limits"),
-    System.ComponentModel.Description("The end value of the range. Must be greater than RangeStartValue.")]
-    public Single EndValue
-    {
-      get { return _EndValue; }
-      set
-      {
-        NotifyChanging();
-        if (Owner != null)
-        {
-          if (value < Owner.MinValue) value = Owner.MinValue;
-          if (value > Owner.MaxValue) value = Owner.MaxValue;
-        }
-        _EndValue = value;
-        NotifyOwner();
-        NotifyChanged();
-      }
-    }
-    private Single _EndValue;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The inner radius of the range.")]
-    public Int32 InnerRadius
-    {
-      get { return _InnerRadius; }
-      set { if (value > 0) { NotifyChanging(); _InnerRadius = value; NotifyOwner(); NotifyChanged(); } }
-    }
-    private Int32 _InnerRadius = 70;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The outer radius of the range.")]
-    public Int32 OuterRadius
-    {
-      get { return _OuterRadius; }
-      set { if (value > 0) { NotifyChanging(); _OuterRadius = value; NotifyOwner(); NotifyChanged(); } }
-    }
-    private Int32 _OuterRadius = 80;
-  }
-  #endregion
-
-  #region [ Gauge Label ]
-  public class AGaugeLabelCollection : CollectionBase
-  {
-    private AGauge Owner;
-    public AGaugeLabelCollection(AGauge sender) { Owner = sender; }
-    private void NotifyChanging()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanging();
-      }
-    }
-
-    private void NotifyChanged()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanged();
-      }
-    }
-
-
-    public AGaugeLabel this[int index] { get { return (AGaugeLabel)List[index]; } }
-    public bool Contains(AGaugeLabel itemType) { return List.Contains(itemType); }
-    public int Add(AGaugeLabel itemType)
-    {
-      itemType.SetOwner(Owner);
-      if (string.IsNullOrEmpty(itemType.Name)) itemType.Name = GetUniqueName();
-      var ret = List.Add(itemType);
-      if (Owner != null) Owner.RepaintControl();
-      return ret;
-    }
-    public void Remove(AGaugeLabel itemType)
-    {
-      List.Remove(itemType);
-      if (Owner != null) Owner.RepaintControl();
-    }
-    public void Insert(int index, AGaugeLabel itemType)
-    {
-      itemType.SetOwner(Owner);
-      if (string.IsNullOrEmpty(itemType.Name)) itemType.Name = GetUniqueName();
-      List.Insert(index, itemType);
-      if (Owner != null) Owner.RepaintControl();
-    }
-    public int IndexOf(AGaugeLabel itemType) { return List.IndexOf(itemType); }
-    public AGaugeLabel FindByName(string name)
-    {
-      foreach (AGaugeLabel ptrRange in List)
-      {
-        if (ptrRange.Name == name) return ptrRange;
-      }
-      return null;
-    }
-
-    protected override void OnInsert(int index, object value)
-    {
-      NotifyChanging();
-      if (string.IsNullOrEmpty(((AGaugeLabel)value).Name)) ((AGaugeLabel)value).Name = GetUniqueName();
-      base.OnInsert(index, value);
-      ((AGaugeLabel)value).SetOwner(Owner);
-      NotifyChanged();
-    }
-    protected override void OnRemove(int index, object value)
-    {
-      NotifyChanging();
-      if (Owner != null) Owner.RepaintControl();
-      NotifyChanged();
-    }
-    protected override void OnClear()
-    {
-      NotifyChanging();
-      if (Owner != null) Owner.RepaintControl();
-      NotifyChanged();
-    }
-
-    private string GetUniqueName()
-    {
-      const string Prefix = "GaugeLabel";
-      int index = 1;
-      while (this.Count != 0)
-      {
-        for (int x = 0; x < this.Count; x++)
-        {
-          if (this[x].Name == (Prefix + index.ToString()))
-            continue;
-          else
-            return Prefix + index.ToString();
-        }
-        index++;
-      };
-      return Prefix + index.ToString();
-    }
-  }
-
-  public class AGaugeLabel
-  {
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Design"),
-    System.ComponentModel.DisplayName("(Name)"),
-    System.ComponentModel.Description("Instance Name.")]
-    public string Name { get { return _Name; } set { NotifyChanging(); _Name = value; NotifyChanged(); } }
-    private string _Name;
-
-    [System.ComponentModel.Browsable(false)]
-    public void SetOwner(AGauge value) { Owner = value; }
-    private AGauge Owner;
-
-    private void NotifyOwner() { if (Owner != null) Owner.RepaintControl(); }
-
-    private void NotifyChanging()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanging();
-      }
-    }
-
-    private void NotifyChanged()
-    {
-      if (Owner != null)
-      {
-        Owner.NotifyChanged();
-      }
-    }
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The color of the caption text.")]
-    public Color Color { get { return _Color; } set { NotifyChanging(); _Color = value; NotifyOwner(); NotifyChanged();  } }
-    private Color _Color = Color.FromKnownColor(KnownColor.WindowText);
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The text of the caption.")]
-    public String Text { get { return _Text; } set { NotifyChanging(); _Text = value; NotifyOwner(); NotifyChanged(); } }
-    private String _Text;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("The position of the caption.")]
-    public Point Position { get { return _Position; } set { NotifyChanging(); _Position = value; NotifyOwner(); NotifyChanged(); } }
-    private Point _Position;
-
-    [System.ComponentModel.Browsable(true),
-    System.ComponentModel.Category("Appearance"),
-    System.ComponentModel.Description("Font of Text.")]
-    public Font Font { get { return _Font; } set { NotifyChanging(); _Font = value; NotifyOwner(); NotifyChanged(); } }
-    private Font _Font = DefaultFont;
-
-    public void ResetFont() { NotifyChanging(); _Font = DefaultFont; NotifyChanged(); }
-    private Boolean ShouldSerializeFont() { return (_Font != DefaultFont); }
-    private static Font DefaultFont = System.Windows.Forms.Control.DefaultFont;
-  }
-  #endregion
-
-  #region [ Gauge Enum ]
-
-  /// <summary>
-  /// First needle color
-  /// </summary>
-  public enum AGaugeNeedleColor
-  {
-    Gray = 0,
-    Red = 1,
-    Green = 2,
-    Blue = 3,
-    Yellow = 4,
-    Violet = 5,
-    Magenta = 6,
-    White = 7,
-  };
-
-  public enum NeedleType
-  {
-    Advance,
-    Simple
-  }
-
-  #endregion
-
-  #region [ EventArgs ]
-  /// <summary>
-  /// Event argument for <see cref="ValueInRangeChanged"/> event.
-  /// </summary>
-  public class ValueInRangeChangedEventArgs : EventArgs
-  {
-    /// <summary>
-    /// Affected GaugeRange
-    /// </summary>
-    public AGaugeRange Range { get; private set; }
-    /// <summary>
-    /// Gauge Value
-    /// </summary>
-    public Single Value { get; private set; }
-    /// <summary>
-    /// True if value is within current range.
-    /// </summary>
-    public bool InRange { get; private set; }
-    public ValueInRangeChangedEventArgs(AGaugeRange range, Single value, bool inRange)
-    {
-      this.Range = range;
-      this.Value = value;
-      this.InRange = inRange;
-    }
-  }
+#region [ EventArgs ]
   #endregion
 
   [System.Runtime.CompilerServices.CompilerGenerated]
